refactor: feature/#50 CommentService to CommentServiceImpl
